HalTheBot/requirements.txt
adal==1.2.7
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiounittest==1.3.0
arrow==1.2.1
async-timeout==3.0.1
attrs==21.2.0
azure-cognitiveservices-language-luis==0.2.0
azure-common==1.1.27
azure-core==1.21.1
azure-cosmos==4.2.0
azure-functions==1.8.0
Babel==2.7.0
binaryornot==0.4.4
botbuilder-ai==4.14.0
botbuilder-core==4.14.0
botbuilder-dialogs==4.14.0
botbuilder-integration-aiohttp==4.14.0
botbuilder-schema==4.14.0
botframework-connector==4.14.0
botframework-streaming==4.14.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.0
cryptography==3.4.8
datedelta==1.3
emoji==1.6.1
frozenlist==1.2.0
future==0.18.2
grapheme==0.6.0
gunicorn==20.1.0
idna==2.10
isodate==0.6.1
Jinja2==3.0.3
jinja2-time==0.2.0
jsonpickle==1.4.2
MarkupSafe==2.0.1
msal==1.6.0
msrest==0.6.10
msrestazure==0.6.4
multidict==5.2.0
multipledispatch==0.6.0
oauthlib==3.1.1
poyo==0.5.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.3
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
whichcraft==0.6.1
yarl==1.4.2
